+++ gutenberg.php
@@ -66,22 +66,6 @@ function gutenberg_menu() {
 			);
 		}
 	}
-
-	if ( gutenberg_is_fse_theme() ) {
-		add_menu_page(
-			__( 'Site Editor (beta)', 'gutenberg' ),
-			sprintf(
-				/* translators: %s: "beta" label. */
-				__( 'Site Editor %s', 'gutenberg' ),
-				'<span class="awaiting-mod">' . __( 'beta', 'gutenberg' ) . '</span>'
-			),
-			'edit_theme_options',
-			'gutenberg-edit-site',
-			'gutenberg_edit_site_page',
-			'dashicons-layout'
-		);
-	}
-
 	if ( current_user_can( 'edit_posts' ) ) {
 		add_submenu_page(
 			'gutenberg',
@@ -109,6 +93,31 @@ function gutenberg_menu() {
 	);
 }
 add_action( 'admin_menu', 'gutenberg_menu', 9 );
+
+/**
+ * Site editor's Menu.
+ *
+ * Adds a new wp-admin menu item for the Site editor.
+ *
+ * @since 9.4.0
+ */
+function gutenberg_site_editor_menu() {
+	if ( gutenberg_is_fse_theme() ) {
+		add_menu_page(
+			__( 'Site Editor (beta)', 'gutenberg' ),
+			sprintf(
+			/* translators: %s: "beta" label. */
+				__( 'Site Editor %s', 'gutenberg' ),
+				'<span class="awaiting-mod">' . __( 'beta', 'gutenberg' ) . '</span>'
+			),
+			'edit_theme_options',
+			'gutenberg-edit-site',
+			'gutenberg_edit_site_page',
+			'dashicons-layout'
+		);
+	}
+}
+add_action( 'admin_menu', 'gutenberg_site_editor_menu', 9 );
 
 /**
  * Modify WP admin bar.
@@ -159,7 +168,7 @@ function gutenberg_build_files_notice() {
  */
 function gutenberg_pre_init() {
 	global $wp_version;
-	if ( defined( 'GUTENBERG_DEVELOPMENT_MODE' ) && GUTENBERG_DEVELOPMENT_MODE && ! file_exists( dirname( __FILE__ ) . '/build/blocks' ) ) {
+	if ( defined( 'GUTENBERG_DEVELOPMENT_MODE' ) && GUTENBERG_DEVELOPMENT_MODE && ! file_exists( __DIR__ . '/build/blocks' ) ) {
 		add_action( 'admin_notices', 'gutenberg_build_files_notice' );
 		return;
 	}
@@ -175,7 +184,7 @@ function gutenberg_pre_init() {
 		return;
 	}
 
-	require_once dirname( __FILE__ ) . '/lib/load.php';
+	require_once __DIR__ . '/lib/load.php';
 }
 
 /**