+++ lib/class-wp-rest-block-types-controller.php
@@ -133,7 +133,7 @@ class WP_REST_Block_Types_Controller extends WP_REST_Controller {
 			$namespace = $request['namespace'];
 		}
 
-		foreach ( $block_types as $slug => $obj ) {
+		foreach ( $block_types as $obj ) {
 			if ( $namespace ) {
 				$pieces          = explode( '/', $obj->name );
 				$block_namespace = $pieces[0];