+++ packages/rich-text/src/component/use-anchor-ref.js
@@ -26,9 +26,9 @@ import { getActiveFormat } from '../get-active-format';
  *
  * @return {Element|Range} The active element or selection range.
  */
-export function useAnchorRef( { ref, value, settings } ) {
+export function useAnchorRef( { ref, value, settings = {} } ) {
 	const { tagName, className, name } = settings;
-	const activeFormat = getActiveFormat( value, name );
+	const activeFormat = name ? getActiveFormat( value, name ) : undefined;
 
 	return useMemo( () => {
 		const { ownerDocument } = ref.current;