+++ packages/react-native-editor/CHANGELOG.md
@@ -11,6 +11,9 @@ For each user feature we should also add a importance categorization label  to i
 
 ## Unreleased
 
+* [***] Adding support for selecting different unit of value in Cover and Columns blocks
+* [**] Button block - Add link picker to the block settings
+
 ## 1.41.0
 
 * [***] Faster editor start and overall operation on Android [#26732]